fix user verification repeat show dialog in service bug
