+++ app/src/main/java/com/nicole/petnanny/ui/profile/pet/PetViewModel.kt
@@ -1,5 +1,6 @@
 package com.nicole.petnanny.ui.profile.pet
 
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -14,6 +15,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 import com.nicole.petnanny.data.Result
+import com.nicole.petnanny.ui.login.UserManager
 
 class PetViewModel(private val repository: PetNannyRepository):ViewModel() {
 
@@ -62,6 +64,7 @@ class PetViewModel(private val repository: PetNannyRepository):ViewModel() {
 //        else{
             getPetsResult()
 //        }
+        getUserResult(UserManager.user.value?.userEmail)
     }
 
 
@@ -96,6 +99,45 @@ class PetViewModel(private val repository: PetNannyRepository):ViewModel() {
                 }
             }
             _refreshStatus.value = false
+        }
+    }
+
+        //    get user result save user verification to userManager
+    fun getUserResult(userEmail: String?) {
+        coroutineScope.launch {
+
+            _status.value = LoadApiStatus.LOADING
+
+            userEmail?.let {
+                val result = repository.getUser(it)
+                Log.d("@@@@", "@@${result} ")
+
+                UserManager.user.value = when (result) {
+                    is Result.Success -> {
+                        _error.value = null
+                        _status.value = LoadApiStatus.DONE
+                        Log.d("@@@@", "@@${result.data} ")
+                        result.data
+                    }
+                    is Result.Fail -> {
+                        _error.value = result.error
+                        _status.value = LoadApiStatus.ERROR
+                        null
+                    }
+                    is Result.Error -> {
+                        _error.value = result.exception.toString()
+                        _status.value = LoadApiStatus.ERROR
+                        null
+                    }
+                    else -> {
+                        _error.value =
+                            PetNannyApplication.instance.getString(R.string.you_know_nothing)
+                        _status.value = LoadApiStatus.ERROR
+                        null
+                    }
+                }
+                _refreshStatus.value = false
+            }
         }
     }
