+++ app/src/main/java/com/nicole/petnanny/ui/profile/service/ServiceViewModel.kt
@@ -62,7 +62,13 @@ class ServiceViewModel(private val repository: PetNannyRepository): ViewModel() 
     }
 
     init {
-        getUserResult(UserManager.user.value?.userEmail)
+        if(UserManager.user.value?.verification == true) {
+                    getServicesResult()
+                    isShowDialog.value = false
+                } else {
+                    isShowDialog.value = true
+                }
+//        getUserResult(UserManager.user.value?.userEmail)
     }
 
     fun getServicesResult() {
@@ -98,50 +104,50 @@ class ServiceViewModel(private val repository: PetNannyRepository): ViewModel() 
         }
     }
 
-    //    get user result save user verification to userManager
-    fun getUserResult(userEmail: String?) {
-        coroutineScope.launch {
-
-            _status.value = LoadApiStatus.LOADING
-
-            userEmail?.let {
-                val result = repository.getUser(it)
-                Log.d("@@@@", "@@${result} ")
-
-                UserManager.user.value = when (result) {
-                    is Result.Success -> {
-                        _error.value = null
-                        _status.value = LoadApiStatus.DONE
-                        Log.d("@@@@", "@@${result.data} ")
-                        result.data
-                    }
-                    is Result.Fail -> {
-                        _error.value = result.error
-                        _status.value = LoadApiStatus.ERROR
-                        null
-                    }
-                    is Result.Error -> {
-                        _error.value = result.exception.toString()
-                        _status.value = LoadApiStatus.ERROR
-                        null
-                    }
-                    else -> {
-                        _error.value =
-                            PetNannyApplication.instance.getString(R.string.you_know_nothing)
-                        _status.value = LoadApiStatus.ERROR
-                        null
-                    }
-                }
-                if(UserManager.user.value?.verification == true) {
-                    getServicesResult()
-                    isShowDialog.value = false
-                } else {
-                    isShowDialog.value = true
-                }
-                _refreshStatus.value = false
-            }
-        }
-    }
+//    //    get user result save user verification to userManager
+//    fun getUserResult(userEmail: String?) {
+//        coroutineScope.launch {
+//
+//            _status.value = LoadApiStatus.LOADING
+//
+//            userEmail?.let {
+//                val result = repository.getUser(it)
+//                Log.d("@@@@", "@@${result} ")
+//
+//                UserManager.user.value = when (result) {
+//                    is Result.Success -> {
+//                        _error.value = null
+//                        _status.value = LoadApiStatus.DONE
+//                        Log.d("@@@@", "@@${result.data} ")
+//                        result.data
+//                    }
+//                    is Result.Fail -> {
+//                        _error.value = result.error
+//                        _status.value = LoadApiStatus.ERROR
+//                        null
+//                    }
+//                    is Result.Error -> {
+//                        _error.value = result.exception.toString()
+//                        _status.value = LoadApiStatus.ERROR
+//                        null
+//                    }
+//                    else -> {
+//                        _error.value =
+//                            PetNannyApplication.instance.getString(R.string.you_know_nothing)
+//                        _status.value = LoadApiStatus.ERROR
+//                        null
+//                    }
+//                }
+//                if(UserManager.user.value?.verification == true) {
+//                    getServicesResult()
+//                    isShowDialog.value = false
+//                } else {
+//                    isShowDialog.value = true
+//                }
+//                _refreshStatus.value = false
+//            }
+//        }
+//    }
 
     fun displayEditServiceDetailsComplete () {
         _navigationToEditSerciveDetail.value = null
